+++ scripts/macros/2014/spells/antagonize.js
@@ -54,7 +54,9 @@ async function use({workflow}) {
             if (nearbyTargets.length === 1) {
                 target = nearbyTargets[0].document;
             } else {
-                [{document: target}, selected] = await dialogUtils.selectTargetDialog(workflow.item.name, 'CHRISPREMADES.Macros.Antagonize.SelectTarget', nearbyTargets);
+                let choice = await dialogUtils.selectTargetDialog(workflow.item.name, 'CHRISPREMADES.Macros.Antagonize.SelectTarget', nearbyTargets);
+                if (!choice) return;
+                [{document: target}, selected] = choice;
                 if (!selected) continue;
             }
             await socketUtils.remoteRollItem(selectedWeapon, {}, {targetUuids: [target.uuid]}, socketUtils.firstOwner(targetToken).id);